backend/requirements.txt
Django==6.0.1
dj-database-url==3.1.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2==2.9.9
sqlparse==0.5.5
tzdata==2025.3
gunicorn==23.0.0
whitenoise==6.9.0
